backend/requirements.txt
# FastAPI and ASGI server
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Environment and configuration
python-dotenv==1.0.0
pydantic-settings==2.0.3

# AI and OpenAI
openai==1.3.7

# Slack integration
slack-sdk==3.26.1

# GitHub integration
PyGithub==1.59.1

# Jira integration
jira==3.5.2

# Google Calendar integration
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0

# HTTP client
httpx==0.25.2

# Data validation and serialization
pydantic==2.5.0

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Logging and monitoring
structlog==23.2.0 